Refactor authentication system
It is now possible to assign many authentication check functions to a RPC method. In addition, the system is more flexible and will allow future updates. The check functions now take a request object as main argument, so it is possible to perform authentication on non-user values (api keys, host, etc.)