Huge tests refactoring
Refactored  tree to split unit & functional tests in separate folders.
Make an intensive use of pytest fixtures (with params) and tests parametrization to increase the number of variations to test.
Used test classes when possible to ease the hot replacement of fixtures for a given set of tests

Total number of lines in test files decreased from 2990 to 2333 while number of actual tests increased a lot, thanks to combinatory