Fix passing of entrypoint to handler
Handler constructor expects to receive entrypoint name
but it was given a protocol by mistake.

One of the end results was that when you don't specify protocol,
Protocol.ALL is used which is the same string as "all entrypoints"
making all methods available in all entrypoints regardless
of what you specify in entrypoint configuration.